Countdown Refactoring
* Margin in the Generic Button has been removed.
* One of the paddings in the Generic Button has been removed.
* GetIt has been added to the project.
* Countdown Counter has been moved under the Countdown Section.
