+++ lib/core/ui/generic_button.dart
@@ -6,13 +6,11 @@ class GenericButton extends StatelessWidget {
     @required this.title,
     @required this.onPressed,
     this.isFilledButton = false,
-    this.margin = const EdgeInsets.symmetric(vertical: 8, horizontal: 4),
     this.padding = EdgeInsets.zero,
     this.textStyle = const TextStyle(),
     Key key,
   })  : assert(title != null),
         assert(onPressed != null),
-        assert(margin != null),
         assert(padding != null),
         assert(textStyle != null),
         super(key: key);
@@ -25,22 +23,18 @@ class GenericButton extends StatelessWidget {
   ///
   /// The background color is `ThemeHelper.appBarActionColor`
   final bool isFilledButton;
-  final EdgeInsets margin;
   final EdgeInsets padding;
   final TextStyle textStyle;
 
   @override
-  Widget build(BuildContext context) => Padding(
-        padding: margin,
-        child: TextButton(
-          style: buildTextButtonStyle,
-          onPressed: onPressed,
-          child: buildButtonTitle,
-        ),
+  Widget build(BuildContext context) => TextButton(
+        style: buildTextButtonStyle,
+        onPressed: onPressed,
+        child: buildButtonTitle,
       );
 
   Widget get buildButtonTitle => Padding(
-        padding: const EdgeInsets.symmetric(vertical: 2, horizontal: 10),
+        padding: padding,
         child: Text(title, style: buildTextStyle),
       );
 
@@ -55,6 +49,5 @@ class GenericButton extends StatelessWidget {
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(8),
         ),
-        padding: padding,
       );
 }
