+++ lib/core/utils/get_it_initializer.dart
@@ -1,0 +1,8 @@
+import 'package:festival_flutterturkiye_org/countdown/logic/countdown_repository.dart';
+import 'package:get_it/get_it.dart';
+
+GetIt getIt = GetIt.instance;
+
+void initializeGetIt() {
+  getIt.registerLazySingleton(() => CountdownRepository());
+}
