+++ lib/countdown/logic/countdown_repository.dart
@@ -4,32 +4,17 @@ final DateTime _eventStartingDate = DateTime(2021, 4, 17, 11);
 final DateTime _eventCompletingDate = DateTime(2021, 4, 18, 20);
 
 class CountdownRepository {
-  factory CountdownRepository() => _instance;
-
-  CountdownRepository._();
-
-  static final CountdownRepository _instance = CountdownRepository._();
-
   Duration get countdownInitialData {
     if (!isEventStarted) {
-      return eventStartingRemainingTime;
+      return _eventStartingRemainingTime;
     } else if (!isEventCompleted) {
-      return eventCompletingRemainingTime;
+      return _eventCompletingRemainingTime;
     }
     return const Duration();
   }
 
-  Duration get eventStartingRemainingTime =>
-      _remainingTimeCalculator(_eventStartingDate);
-
-  Duration get eventCompletingRemainingTime =>
-      _remainingTimeCalculator(_eventCompletingDate);
-
-  Duration _remainingTimeCalculator(DateTime dateTime) =>
-      dateTime.difference(DateTime.now());
-
-  bool get isEventStarted => eventStartingRemainingTime.isNegative;
-  bool get isEventCompleted => eventCompletingRemainingTime.isNegative;
+  bool get isEventStarted => _eventStartingRemainingTime.isNegative;
+  bool get isEventCompleted => _eventCompletingRemainingTime.isNegative;
 
   EventStatus get eventStatus {
     if (isEventCompleted) {
@@ -43,22 +28,25 @@ class CountdownRepository {
 
   Stream<Duration> getCountdown() {
     if (!isEventStarted) {
-      return getStartingCountdown();
+      return _countdownDurationCalculator(_eventStartingDate);
     } else if (!isEventCompleted) {
-      return getCompletingCountdown();
+      return _countdownDurationCalculator(_eventCompletingDate);
     }
     return const Stream.empty();
   }
 
-  Stream<Duration> getStartingCountdown() =>
-      _countdownDurationCalculator(_eventStartingDate);
+  Duration get _eventStartingRemainingTime =>
+      _remainingTimeCalculator(_eventStartingDate);
 
-  Stream<Duration> getCompletingCountdown() =>
-      _countdownDurationCalculator(_eventCompletingDate);
+  Duration get _eventCompletingRemainingTime =>
+      _remainingTimeCalculator(_eventCompletingDate);
 
   Stream<Duration> _countdownDurationCalculator(DateTime dateTime) =>
       Stream.periodic(
         const Duration(seconds: 1),
         (_) => _remainingTimeCalculator(dateTime),
       );
+
+  Duration _remainingTimeCalculator(DateTime dateTime) =>
+      dateTime.difference(DateTime.now());
 }
