+++ lib/countdown/ui/countdown_section.dart
@@ -1,3 +1,4 @@
+import 'package:festival_flutterturkiye_org/core/utils/get_it_initializer.dart';
 import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
 
@@ -6,7 +7,6 @@ import 'package:festival_flutterturkiye_org/core/utils/responsive_helper.dart';
 import 'package:festival_flutterturkiye_org/core/utils/theme_helper.dart';
 import 'package:festival_flutterturkiye_org/countdown/logic/countdown_repository.dart';
 import 'package:festival_flutterturkiye_org/countdown/model/event_status.dart';
-import 'package:festival_flutterturkiye_org/countdown/ui/countdown_counter.dart';
 import 'package:festival_flutterturkiye_org/countdown/ui/countdown_text_builder.dart';
 import 'package:festival_flutterturkiye_org/countdown/ui/countdown_text_span.dart';
 import 'package:festival_flutterturkiye_org/widgets/responsive_builder.dart';
@@ -82,28 +82,28 @@ class _CountdownSectionState extends State<CountdownSection> {
               children: [
                 _CountdownTitle(),
                 SizedBox(height: screenSize.height * 0.1),
-                CountdownCounter(),
+                _CountdownCounter(),
                 SizedBox(height: screenSize.height * 0.1),
                 const Center(
                   child: ResponsiveBuilder(
                     smallWidget: _RegisterButton(
                       fontSize: 32,
                       padding: EdgeInsets.symmetric(
-                        vertical: 8,
-                        horizontal: 16,
+                        vertical: 4,
+                        horizontal: 8,
                       ),
                     ),
                     mediumWidget: _RegisterButton(
                       fontSize: 40,
                       padding: EdgeInsets.symmetric(
-                        vertical: 12,
-                        horizontal: 20,
+                        vertical: 8,
+                        horizontal: 16,
                       ),
                     ),
                     largeWidget: _RegisterButton(
                       fontSize: 48,
                       padding: EdgeInsets.symmetric(
-                        vertical: 16,
+                        vertical: 12,
                         horizontal: 24,
                       ),
                     ),
@@ -129,7 +129,7 @@ class _CountdownSectionState extends State<CountdownSection> {
 }
 
 class _CountdownTitle extends StatelessWidget {
-  _CountdownTitle() : _countdownRepository = CountdownRepository();
+  _CountdownTitle() : _countdownRepository = getIt.get<CountdownRepository>();
 
   final CountdownRepository _countdownRepository;
 
@@ -181,4 +181,83 @@ class _RegisterButton extends StatelessWidget {
         textStyle: TextStyle(fontSize: fontSize),
         padding: padding,
       );
+}
+
+class _CountdownCounter extends StatelessWidget {
+  _CountdownCounter({Key key})
+      : _countdownRepository = getIt.get<CountdownRepository>(),
+        super(key: key);
+
+  final CountdownRepository _countdownRepository;
+
+  @override
+  Widget build(BuildContext context) => StreamBuilder<Duration>(
+      stream: _countdownRepository.getCountdown(),
+      initialData: _countdownRepository.countdownInitialData,
+      builder: (context, remainingTimeSnapshot) {
+        if (remainingTimeSnapshot.hasData) {
+          final remainingTime = remainingTimeSnapshot.data;
+          var textSpans = <TextSpan>[];
+
+          switch (_countdownRepository.eventStatus) {
+            case EventStatus.completed:
+              textSpans = _completedTextSpan(remainingTime);
+              break;
+            case EventStatus.started:
+              textSpans = _startedTextSpan(remainingTime);
+              break;
+            case EventStatus.waiting:
+              textSpans = _waitingTextSpan(remainingTime);
+              break;
+          }
+
+          return CountdownTextBuilder(textSpans: textSpans);
+        }
+        return const SizedBox.shrink();
+      });
+
+  String _dateFixer(int date) => '${date < 10 ? '0' : ''}$date';
+
+  List<TextSpan> _waitingTextSpan(Duration remainingTime) => <TextSpan>[
+        _DateText(_dateFixer(remainingTime.inDays)),
+        CountdownTextSpan(' gün '),
+        _DateText(_dateFixer(remainingTime.inHours % 24)),
+        CountdownTextSpan(' saat '),
+        _DateText(_dateFixer(remainingTime.inMinutes % 60)),
+        CountdownTextSpan(' dakika '),
+        _DateText(_dateFixer(remainingTime.inSeconds % 60)),
+        CountdownTextSpan(' saniye kaldı'),
+      ];
+
+  List<TextSpan> _startedTextSpan(Duration remainingTime) => <TextSpan>[
+        _DateText(_dateFixer(remainingTime.inDays)),
+        CountdownTextSpan(' gün '),
+        _DateText(_dateFixer(remainingTime.inHours % 24)),
+        CountdownTextSpan(' saat '),
+        _DateText(_dateFixer(remainingTime.inMinutes % 60)),
+        CountdownTextSpan(' dakika '),
+        _DateText(_dateFixer(remainingTime.inSeconds % 60)),
+        CountdownTextSpan(' saniye kaldı'),
+      ];
+
+  List<TextSpan> _completedTextSpan(Duration remainingTime) => <TextSpan>[
+        CountdownTextSpan(
+          'Etkinliğimize göstermiş olduğunuz ilgi için\n',
+        ),
+        CountdownTextSpan(
+          'TEŞEKKÜR EDERİZ',
+          fontWeight: FontWeight.bold,
+        )
+      ];
+}
+
+/// The only difference with [CountdownTextSpan] is [FontWeight]
+///
+/// Default [FontWeight] is `FontWeight.bold`
+class _DateText extends CountdownTextSpan {
+  /// If you give a [fontWeight], it will override the default [FontWeight].
+  _DateText(String text, {FontWeight fontWeight = FontWeight.bold})
+      : assert(text != null),
+        assert(fontWeight != null),
+        super(text, fontWeight: fontWeight);
 }
