+++ lib/main.dart
@@ -1,8 +1,10 @@
+import 'package:festival_flutterturkiye_org/core/utils/get_it_initializer.dart';
 import 'package:festival_flutterturkiye_org/core/utils/theme_helper.dart';
 import 'package:festival_flutterturkiye_org/pages/home_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
+  initializeGetIt();
   runApp(const FlutterFestivalApp());
 }
