+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
+  get_it: ^6.0.0
   material_design_icons_flutter: ^4.0.5855
   url_launcher: ^5.7.10
   video_player: ^1.0.1
